+++ packages/xo-server/src/xo-mixins/plugins.mjs
@@ -20,12 +20,12 @@ export default class {
     }).addVocabulary(['$multiline', '$type', 'enumNames'])
     this._plugins = { __proto__: null }
 
-    this._pluginsMetadata = new PluginsMetadata({
-      connection: app._redis,
-      namespace: 'plugin-metadata',
-    })
-
     app.hooks.on('start', () => {
+      this._pluginsMetadata = new PluginsMetadata({
+        connection: app._redis,
+        namespace: 'plugin-metadata',
+      })
+
       app.addConfigManager(
         'plugins',
         () => this._pluginsMetadata.get(),